requirements.txt
Flask==2.0.1
torch==2.0.0
torchvision==0.15.0
Pillow==8.2.0
gdown